# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MathAgent 研究状态服务",
    "description": "本地研究工作台与受控运行接口。support 表示证据策略下的支持，不是数学真值。Fake 运行均是确定性测试。",
    "version": "0.2.0"
  },
  "paths": {
...
    "/projects/{project_id}/branches": {
      "get": {
        "tags": [
          "workspace"
        ],
        "summary": "Branches",
        "operationId": "branches_projects__project_id__branches_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/branches/{branch_id}/layout": {
      "get": {
        "tags": [
          "workspace"
        ],
        "summary": "Layout",
        "operationId": "layout_branches__branch_id__layout_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "branch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Branch Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "workspace"
        ],
        "summary": "Update Layout",
        "operationId": "update_layout_branches__branch_id__layout_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "branch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Branch Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 200,
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LayoutUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/annotations": {
      "post": {
        "tags": [
          "workspace"
        ],
        "summary": "Annotation",
        "operationId": "annotation_annotations_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 200,
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnnotationCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/branches/{branch_id}/annotations": {
      "get": {
        "tags": [
          "workspace"
        ],
        "summary": "Annotations",
        "operationId": "annotations_branches__branch_id__annotations_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "branch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Branch Id"
            }
          },
          {
            "name": "revision_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Revision Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/manuscript/blocks/{block_id}/revisions": {
      "post": {
        "tags": [
          "workspace"
        ],
        "summary": "Update Block",
        "operationId": "update_block_manuscript_blocks__block_id__revisions_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "block_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Block Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 200,
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BlockUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "workspace"
        ],
        "summary": "Block Revisions",
        "operationId": "block_revisions_manuscript_blocks__block_id__revisions_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "block_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Block Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/conflicts/{conflict_id}/resolve": {
      "post": {
        "tags": [
          "workspace"
        ],
        "summary": "Resolve",
        "operationId": "resolve_conflicts__conflict_id__resolve_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "conflict_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Conflict Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 200,
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveConflict"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{project_id}/search": {
      "get": {
        "tags": [
          "workspace"
        ],
        "summary": "Search",
        "operationId": "search_projects__project_id__search_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 300,
              "title": "Q"
            }
          },
          {
            "name": "branch_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Branch Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AnnotationCreate": {
        "properties": {
          "branch_id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Branch Id"
          },
          "revision_id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Revision Id"
          },
          "body": {
            "type": "string",
            "maxLength": 200000,
            "minLength": 1,
            "title": "Body"
          },
          "anchor_quote": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20000,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Anchor Quote"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "branch_id",
          "revision_id",
          "body"
        ],
        "title": "AnnotationCreate"
      },
...
      "BlockUpdate": {
        "properties": {
          "branch_id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Branch Id"
          },
          "expected_version": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Expected Version"
          },
          "expected_body": {
            "type": "string",
            "maxLength": 200000,
            "title": "Expected Body"
          },
          "body": {
            "type": "string",
            "maxLength": 200000,
            "title": "Body"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "branch_id",
          "expected_version",
          "expected_body",
          "body"
        ],
        "title": "BlockUpdate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LayoutUpdate": {
        "properties": {
          "expected_version": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Expected Version"
          },
          "positions": {
            "additionalProperties": {
              "$ref": "#/components/schemas/Position"
            },
            "propertyNames": {
              "maxLength": 100,
              "minLength": 1
            },
            "type": "object",
            "maxProperties": 10000,
            "title": "Positions"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expected_version",
          "positions"
        ],
        "title": "LayoutUpdate"
      },
...
      "Position": {
        "properties": {
          "x": {
            "type": "number",
            "maximum": 1000000.0,
            "minimum": -1000000.0,
            "title": "X"
          },
          "y": {
            "type": "number",
            "maximum": 1000000.0,
            "minimum": -1000000.0,
            "title": "Y"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "x",
          "y"
        ],
        "title": "Position"
      },
...
      "ResolveConflict": {
        "properties": {
          "branch_id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Branch Id"
          },
          "expected_current_revision_id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Expected Current Revision Id"
          },
          "selected_revision_id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Selected Revision Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "branch_id",
          "expected_current_revision_id",
          "selected_revision_id"
        ],
        "title": "ResolveConflict"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}